+++ notebooks/ORR.jl
@@ -104,7 +104,7 @@ begin
     const nref = 0
     const κ = 10.0
     const vfac = 1
-    const flux = LiquidElectrolytes.sflux!
+    const flux! = LiquidElectrolytes.sflux!
     const R0 = 5.0e-16mol / (cm^2 * s)
 
     const Δg = 0.0
@@ -190,7 +190,7 @@ begin
         Γ_we = 1,
         Γ_bulk = 2,
         eneutral = false,
-        flux,
+        flux!,
         rlog = mylog
     )