+++ CHANGELOG.md
@@ -5,15 +5,17 @@ All notable changes to this project will be documented in this file.
 ## 2.0.0 
 
 ### Breaking
-- `aflux`, `sflux` and `cflux` replaced with  `aflux!`, `sflux!` and `cflux!` 
-- `scheme` paramter of ElectrolyteData replaced by `flux` parameter which should be one of `aflux!`, `sflux!` and `cflux!` 
+- `aflux`, `sflux` and `cflux` calculating just one activity coefficinet
+  replaced with  `aflux!`, `sflux!` and `cflux!`  calculating all activity coefficients at once
+- `scheme` paramter of ElectrolyteData replaced by `flux!` parameter which should be one of `aflux!`, `sflux!` and `cflux!` 
 - PNPSystem and PBSystem are now distinct types in order to be able to implement more specific methods
 - pnpunkonowns has been removed in favor of method of VoronoiFVM.unknowns
 - rexp, rlog are now part of ElectrolyteData
 
 ### Features
-- ElectrolyteData now has a function entry `γ!` for user definable function which calculates activity coefficients
+- ElectrolyteData now has a function entry `actcoeff!` for user definable function which calculates activity coefficients
 - Fixed implementation of PBSystem, now cases with different `κ` etc should work
+- Pre-calculate more data in ElectrolyteData, use `update_derived!` to update these fields.
 
 
 ## 1.1.0 - 2025-03-16